Update grails-datastore-mongo/src/main/groovy/org/grails/datastore/mapping/mongo/engine/MongoEntityPersister.java
There's no check that the 'previous' object being passed in is not null.

I've been getting errors on this for months now, and figured sending a patch back 
might help raise this to someone's attention.

The fact that something is passing in a null might be a bigger issue, 
or it may not be - I'm not sure.  But null checking needs to happen here, 
and possible on line 607 (the ea.getProperty) as well.