Made testing example a full integration test
I had trouble following the example provided in the documentation.

One particular issue was that imports were missing, so I had to go back to the gorm tests to see the static imports actually being used to refer to requestTo.

The suggested changes work with the provided example in the other sections of the manual.