GRAILS-12039 - improve validation
This fix is likely temporary.  We have an ordering problem where the validator in GormValidationApi is initialized with null because it is being initialized with the validator from the GrailsDomainClass but the validator in the GrailsDomainClass isn't initialized until later in the process.  This change makes the problem go away for now, at least for Hibernate.  I am working on a better solution.