Use @CompileDynamic for JPAQL query methods. Fixes #924
Hibernate 5.1 and 5.2 are not binary compatible when it comes to the createQuery method since they return different types (org.hibernate.Query for 5.1 and org.hibernate.query.Query for 5.2+). The only way to resolve this is to use dynamic dispatch for the JPA-QL methods by adding @CompileDynamic to all of these methods.