Check if the associated object has already been validated
It is possible for associatedObject to already have been validated
and therefore, without this check, we end up re-validating it and
potentially creating an OutOfMemory error due to nested paths when
creating/adding the errors.
This is only an issue when there are errors present, when there are
no errors the re-validation is not a problem, other than wasting time.